Avoid overwrite of self._topic_name
The line 'self._topic_name = None' was previously placed after
the call to self._subscribe_topic, so _topic_name would be set
based on the parameter server and then deleted right after.
This puts the call to self._subscribe_topic after the line
that resets _topic_name, to avoid that.